get invoices phase 2 14/3/2023

+++ lib/app/models/invoice_details.dart
@@ -11,10 +11,10 @@ class InvoiceDetails {
   });
 
 
-  int? pricePerUnit;
+  double? pricePerUnit;
   String? invoiceId;
-  int? baseAmount;
-  int? quantity;
+  double? baseAmount;
+  double? quantity;
   String? productName;
   String? transactionCurrencyId;
   String? invoiceDetailsId;
